- update the delete method for the products to make sure you are not able to delete the DEFAULT IMAGE

+++ src/main/java/com/weekendwarriors/weekend_warriors_backend/service/ProductService.java
@@ -108,7 +108,7 @@ public class ProductService {
 
     public void deleteProduct(String id) throws IOException {
         Product product = productRepository.findById(id).orElseThrow(() -> new RuntimeException("Product not found with id: " + id));
-        if(product.getImageId() != null)
+        if(product.getImageId() != null && product.getImageId() != imageManagement.provideDefaultImageId())
             imageManagement.deleteImage(product.getImageId());
         productRepository.deleteById(id);
     }
